Set the expire timer
This sets the expire timer of a message such that the current expiration
timer of a conversation stays the same (except in cases where it was
explicitly set to another value).

It may be a little counterproductive when sending a message to (e.g.) a
contact to query the timer for the thread which unwraps it back to
querying the contact, but this method may potentially also be used to
display the expiration timer.

Note that due to #171, I would guess that the expiration timer always
stays the same for contacts directly after linking, e.g. changing it
on an official client and then sending a message from presage, presage
will set the expiration timer back to before the change.

This fixes #93.